+++ src/Exception/InvalidArgumentException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace IndexZer0\HaRestApiClient\Exception;
+
+use InvalidArgumentException as BaseInvalidArgumentException;
+
+class InvalidArgumentException extends BaseInvalidArgumentException implements HaExceptionInterface
+{
+}