+++ CHANGELOG.md
@@ -2,3 +2,20 @@
 
 All notable changes to `ha-rest-api-client` will be documented in this file.
 
+## v2.0.0
+### Added
+- `HaWebhookClient`.
+- Various exception types.
+### Changed
+- Updated documentation in README.md.
+- No longer depend on `guzzlehttp/guzzle` (Guzzle7), instead use HTTPlug for client interoperability.
+- Handle converting http responses to exceptions manually.
+
+## v1.0.1
+### Changed
+- Nothing notable to package function, just code clean up.
+
+## v1.0.0
+- Initial release.
+- `guzzlehttp/guzzle` (Guzzle7) dependency.
+- `HaRestApiClient`.